requirements.txt
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.8
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==1.3.0
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
more-itertools==8.0.2
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.3.1
pytest-django==3.7.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
